+++ lib/routes.dart
@@ -430,6 +430,8 @@ class AppRouterDelegate extends RouterDelegate<RouteConfiguration>
                   graphQlProvider,
                   userRepository,
                   Get.find(),
+                  settingsRepository,
+                  me: me,
                 ),
               );
               AbstractChatRepository chatRepository =
@@ -471,7 +473,6 @@ class AppRouterDelegate extends RouterDelegate<RouteConfiguration>
               deps.put(CallService(
                 Get.find(),
                 chatService,
-                settingsRepository,
                 callRepository,
               ));
 
@@ -529,6 +530,8 @@ class AppRouterDelegate extends RouterDelegate<RouteConfiguration>
                 graphQlProvider,
                 userRepository,
                 Get.find(),
+                settingsRepository,
+                me: me,
               ),
             );
             AbstractChatRepository chatRepository =
@@ -571,7 +574,6 @@ class AppRouterDelegate extends RouterDelegate<RouteConfiguration>
             CallService callService = deps.put(CallService(
               Get.find(),
               chatService,
-              settingsRepository,
               callRepository,
             ));